add: component folder: Card
