+++ src/components/Card/index.jsx
@@ -1,0 +1,5 @@
+import DataSectionLabelCard from './DataSectionLabelCard'
+
+export {
+  DataSectionLabelCard
+}
