+++ src/components/index.js
@@ -1,9 +1,11 @@
 import Dialog from './Dialog/index.jsx'
 import Header from './Header/index.jsx'
 import Sidebar from './Sidebar/index.jsx'
+import Card from './Card/index.jsx'
 
 export {
   Dialog,
   Header,
-  Sidebar
+  Card,
+  Sidebar,
 }
